HW-1 Items model data class 별도로 생성
