+++ class13/Sangmeebee/MyProject/app/src/main/java/com/example/myproject/retrofit/model/Items.kt
@@ -1,0 +1,23 @@
+package com.example.myproject.retrofit.model
+
+import com.google.gson.annotations.SerializedName
+
+data class Items(
+
+    @SerializedName("title")
+    val title: String,
+    @SerializedName("link")
+    val link: String,
+    @SerializedName("image")
+    val image: String,
+    @SerializedName("subtitle")
+    val subtitle: String,
+    @SerializedName("pubDate")
+    val pubDate: Int,
+    @SerializedName("director")
+    val director: String,
+    @SerializedName("actor")
+    val actor: String,
+    @SerializedName("userRating")
+    val userRating: Double
+)
