+++ class13/Sangmeebee/MyProject/app/src/main/java/com/example/myproject/retrofit/model/Movie.kt
@@ -15,23 +15,3 @@ data class Movie(
     @SerializedName("items")
     val items: List<Items>
 )
-
-data class Items(
-
-    @SerializedName("title")
-    val title: String,
-    @SerializedName("link")
-    val link: String,
-    @SerializedName("image")
-    val image: String,
-    @SerializedName("subtitle")
-    val subtitle: String,
-    @SerializedName("pubDate")
-    val pubDate: Int,
-    @SerializedName("director")
-    val director: String,
-    @SerializedName("actor")
-    val actor: String,
-    @SerializedName("userRating")
-    val userRating: Double
-)
